pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docarray>=0.41.0",
    "grandalf>=0.8",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "pydantic>=2.12.5",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
